changed jump and jump if false to take absolute positions instead of offsets, increased the bytecode size to 2 bytes because the jump indexes were larger than a byte and was causing a problem with bigger programs, also partially implmented brainfuck

+++ src_bytecode/Function.hpp
@@ -6,7 +6,7 @@
 #include <string>
 #include <cstdint>
 
-typedef int8_t CODE_SIZE; // Bytecode size, only 8 bits for pointers will be too small for large programs
+typedef int16_t CODE_SIZE; // Bytecode size, only 8 bits for pointers will be too small for large programs
 
 std::string CODE_TO_NUMBER_STRING(CODE_SIZE code){
     return std::to_string((int)code);
